backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
alembic==1.14.*
pydantic-settings==2.7.*
itsdangerous==2.2.*
python-multipart==0.0.*
aiosqlite==0.21.*
asyncpg==0.30.*
passlib[bcrypt]==1.7.*
pyyaml==6.*
aiosmtplib==3.*
pytest==8.*
pytest-asyncio==0.25.*
httpx==0.28.*
